Cpg132 issue fix rajkumar

See merge request IvyProjects/IvyCpg_Product_MobileApp_Android!340

+++ ivycpg/src/main/java/com/ivy/sd/png/model/TaxInterface.java
@@ -43,7 +43,7 @@ public interface TaxInterface {
     HashMap<String, ArrayList<TaxBO>> getmTaxBoBatchProduct();
     ArrayList<TaxBO> getBillTaxList();
     LinkedHashMap<Integer, HashSet<Double>> getTaxPercentagerListByGroupId();
-    float updateProductWiseIncludeTax(List<ProductMasterBO> productMasterBOS);
+    double updateProductWiseIncludeTax(List<ProductMasterBO> productMasterBOS);
     float getTaxAmountByProduct(ProductMasterBO productMasterBO);
